Merge use-site targeted annotations into corresponding Annotations
Add PropertyDescriptor.backingField/delegateField to store annotations
on the field directly in an otherwise almost empty descriptor instance,
instead of storing them with use-sites in the corresponding property
descriptor. Instead of AnnotationWithTarget, create AnnotationDescriptor
instances in AnnotationSplitter. Change DescriptorRenderer to render
annotations on "related" declarations when needed, with the explicit
use-site target if applicable.

Most changes in diagnostic test data are related to the fact that
annotations which are known to have an incompatible use-site to the
declaration they're applied at (such as `@param:`-annotation on a
class), are now not loaded at all. Some changes are also explained by
the fact that for example an annotation on the property which is only
applicable to FIELD is now rendered with an explicit use-site target
`@field:`, regardless of whether it did have that use-site target
syntactically or not.

Basically, after this change there's no point in calling
Annotations.getUseSiteTargetedAnnotations/getAllAnnotations anymore
because it's easier and more intuitive to just use Annotations of the
corresponding object -- the backing / delegate field (introduced in this
commit) or the extension receiver / setter parameter (introduced in
previous commits). Usages of use-site-target-related methods will be
refactored out in subsequent commits